task-api-service/pyproject.toml
[project]
name = "task-api-service"
version = "0.1.0"
description = ""
authors = [
    {name = "sagarapidev",email = "sagar.api.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
